Split mojo, remove deprecated parameter and unify test-handling
This prepare for other test implementations and cleanup of
inconsistencies.
